Adds support for JSON 1.0
Bumps Project.toml to support JSON 1.0. Needed a minor relaxation of the type that is passed into the validate_spec function for tests to pass.

While the tests pass, I notice that there are "duplicate key detected in  mapping" error messages that appear in the output as well. These errors are unrelated to bumping JSON to 1.0 as they appear with the previous version as well.